moved vendor services to new controller, followed separation of concerns pattern with other related files
