+++ UnitTests/Controllers/VendorsControllerShould.cs
@@ -194,160 +194,5 @@ namespace UnitTests.Controllers
             Assert.True(usersResult.Value);
         }
 
-        [Fact]
-        public void GetVendorTypes_ReturnsServiceList()
-        {
-            // arrange
-            var serviceTypesMock = new Mock<VendorServiceTypes>().Object;
-            var serviceTypes = serviceTypesMock.GetVendorServiceTypes();
-
-            // act
-            var task = _sut.GetVendorServiceTypes();
-
-            // assert
-            Assert.IsType<OkObjectResult>(task.Result);
-
-            var result = task.Result as OkObjectResult;
-            var usersResult = result.Value as List<string>;
-            Assert.Equal(serviceTypes, usersResult);
-        }
-
-        [Fact]
-        public void GetVendorsByService_ReturnsVendorsWithService()
-        {
-            // arrange
-            var vendor = new Vendor { id = 123, userName = "vendor@example.com", name = "name1", website = "website_1" };
-            var service = new VendorServices
-            {
-                vendorId = 123,
-                flatFee = true,
-                price = 20,
-                serviceDescription = "desc",
-                serviceName = "svcName",
-                serviceType = "Venue"
-            };
-            vendor.services = new List<VendorServices> { service };
-            var vendors = new List<Vendor> { vendor, vendor, vendor };
-
-            _vendorsQueryMock.Setup(x => x.GetVendorsByServiceTypes("Venue"))
-                .Returns(Task.Factory.StartNew(() => vendors));
-
-            // act
-            var task = _sut.GetVendorsByServiceType("Venue");
-
-            // assert
-            Assert.IsType<OkObjectResult>(task.Result);
-
-            var result = task.Result as OkObjectResult;
-            var usersResult = result.Value as List<Vendor>;
-            Assert.Equal(vendors[2].id, usersResult[2].id);
-            Assert.Equal("Venue", usersResult[0].services.FirstOrDefault().serviceType);
-        }
-
-        [Fact]
-        public void InsertServiceType_ReturnsNewServiceType()
-        {
-            // arrange
-            var service = new VendorServices
-            {
-                vendorId = 123,
-                flatFee = true,
-                price = 20,
-                serviceDescription = "desc",
-                serviceName = "svcName",
-                serviceType = "Venue"
-            };
-           
-            _vendorsQueryMock.Setup(x => x.InsertService(service))
-                .Returns(Task.Factory.StartNew(() => service));
-
-            // act
-            var task = _sut.InsertService(service);
-
-            // assert
-            Assert.IsType<OkObjectResult>(task.Result);
-
-            var result = task.Result as OkObjectResult;
-            var usersResult = result.Value as VendorServices;
-            Assert.Equal(service, usersResult);
-        }
-
-        [Fact]
-        public void UpdateServiceType_ReturnsUpdatedServiceType()
-        {
-            // arrange
-            var service = new VendorServices
-            {
-                vendorId = 123,
-                flatFee = true,
-                price = 20,
-                serviceDescription = "desc",
-                serviceName = "svcName",
-                serviceType = "Venue"
-            };
-
-            _vendorsQueryMock.Setup(x => x.UpdateService(service))
-                .Returns(Task.Factory.StartNew(() => service));
-
-            // act
-            var task = _sut.UpdateService(service);
-
-            // assert
-            Assert.IsType<OkObjectResult>(task.Result);
-
-            var result = task.Result as OkObjectResult;
-            var usersResult = result.Value as VendorServices;
-            Assert.Equal(service, usersResult);
-        }
-
-        [Fact]
-        public void GetServicesByVendor_ReturnsServicesOfVendor()
-        {
-            // arrange
-            var vendor = new Vendor { id = 123, userName = "vendor@example.com", name = "name1", website = "website_1" };
-            var service = new VendorServices
-            {
-                vendorId = 123,
-                flatFee = true,
-                price = 20,
-                serviceDescription = "desc",
-                serviceName = "svcName",
-                serviceType = "Venue"
-            };
-            vendor.services = new List<VendorServices> { service };
-           
-            _vendorsQueryMock.Setup(x => x.GetServicesByVendor(vendor.id.Value))
-                .Returns(Task.Factory.StartNew(() => vendor.services));
-
-            // act
-            var task = _sut.GetServicesByVendor(vendor.id.Value);
-
-            // assert
-            Assert.IsType<OkObjectResult>(task.Result);
-
-            var result = task.Result as OkObjectResult;
-            var usersResult = result.Value as List<VendorServices>;
-            Assert.Equal(vendor.services, usersResult);
-        }
-
-        [Fact]
-        public void DeactivateService_ReturnsTrue()
-        {
-            // arrange
-            var service = new VendorServices { id = 123 };
-
-            _vendorsQueryMock.Setup(x => x.DeactivateService(service.id))
-                .Returns(Task.Factory.StartNew(() => true));
-
-            // act
-            var task = _sut.DeactivateService(service.id);
-
-            // assert
-            Assert.IsType<OkObjectResult>(task.Result);
-
-            var result = task.Result as OkObjectResult;
-            var usersResult = result.Value as bool?;
-            Assert.True(usersResult);
-        }
     }
 }
