+++ source/Controllers/VendorsController.cs
@@ -196,155 +196,5 @@ namespace source.Controllers
                 return new BadRequestResult();
             }
         }
-
-        /// <summary>
-        /// Gets the list of service types for vendors' services
-        /// </summary>
-        /// <returns>List of vendor service types</returns>
-        [HttpGet("servicetypes")]
-        public async Task<IActionResult> GetVendorServiceTypes()
-        {
-            try
-            {
-                VendorServiceTypes types = new VendorServiceTypes();
-                return new OkObjectResult(types.GetVendorServiceTypes());
-            }
-            catch (Exception ex)
-            {
-                await _logger.LogError(HttpContext.User, ex);
-                return new BadRequestResult();
-            }
-        }
-
-        /// <summary>
-        /// Gets list of vendors by service type
-        /// </summary>
-        /// <param name="service">Service type</param>
-        /// <returns>List of vendors with requested service type</returns>
-        [HttpGet("byservice/{service}")]
-        public async Task<ActionResult> GetVendorsByServiceType(string service)
-        {
-            try
-            {
-                List<Vendor> vendors = new List<Vendor>();
-                vendors = await _vendorQuery.GetVendorsByServiceTypes(service);
-
-                if (vendors == null)
-                {
-                    return new NotFoundResult();
-                }
-
-                return new OkObjectResult(vendors);
-            }
-            catch (Exception ex)
-            {
-                await _logger.LogError(HttpContext.User, ex);
-                return new BadRequestResult();
-            }
-        }
-
-        /// <summary>
-        /// Inserts a new vendor service
-        /// </summary>
-        /// <param name="service">The new vendor service</param>
-        /// <returns>Saved vendor service</returns>
-        [HttpPost("service")]
-        public async Task<IActionResult> InsertService([FromBody]VendorServices service)
-        {
-            try
-            {
-                VendorServices newService = await _vendorQuery.InsertService(service);
-                if (newService == null)
-                {
-                    return new NotFoundResult();
-                }
-
-                return new OkObjectResult(newService);
-            }
-            catch (Exception ex)
-            {
-                await _logger.LogError(HttpContext.User, ex);
-                return new BadRequestResult();
-            }
-        }
-
-        /// <summary>
-        /// Updates a vendors service
-        /// </summary>
-        /// <param name="service">Service to be updated</param>
-        /// <returns>Saved vendor service</returns>
-        [HttpPut("service")]
-        public async Task<IActionResult> UpdateService([FromBody]VendorServices service)
-        {
-            try
-            {
-                VendorServices newService = await _vendorQuery.UpdateService(service);
-                if (newService == null)
-                {
-                    return new NotFoundResult();
-                }
-
-                return new OkObjectResult(newService);
-            }
-            catch (Exception ex)
-            {
-                await _logger.LogError(HttpContext.User, ex);
-                return new BadRequestResult();
-            }
-        }
-
-        /// <summary>
-        /// Gets services by vendor
-        /// </summary>
-        /// <param name="id">Vendor id</param>
-        /// <returns>List of vendor services</returns>
-        [HttpGet("servicesbyvendor/{id}")]
-        public async Task<ActionResult> GetServicesByVendor(int id)
-        {
-            try
-            {
-                List<VendorServices> services = new List<VendorServices>();
-                services = await _vendorQuery.GetServicesByVendor(id);
-
-                if (services == null)
-                {
-                    return new NotFoundResult();
-                }
-
-                return new OkObjectResult(services);
-            }
-            catch (Exception ex)
-            {
-                await _logger.LogError(HttpContext.User, ex);
-                return new BadRequestResult();
-            }
-        }
-
-        /// <summary>
-        /// Deactivate a service
-        /// </summary>
-        /// <param name="id">Service id</param>
-        /// <returns>True/False</returns>
-        [HttpPut("deactivateservice/{id}")]
-        public async Task<ActionResult> DeactivateService(int id)
-        {
-            try
-            {
-                bool result = await _vendorQuery.DeactivateService(id);
-
-                if (result == false)
-                {
-                    return new NotFoundResult();
-                }
-
-                return new OkObjectResult(result);
-            }
-            catch (Exception ex)
-            {
-                await _logger.LogError(HttpContext.User, ex);
-                return new BadRequestResult();
-            }
-        }
     }
-
 }
